+++ packages/react/src/edge.ts
@@ -11,14 +11,34 @@ import type {
   prerender as Prerender,
   PrerenderOptions,
 } from 'react-dom/static';
+import { ReactNode } from 'react';
 
 const prerender: typeof Prerender = _prerender;
 const renderToReadableStream: typeof RenderToReadableStream =
   _renderToReadableStream;
 
+type RenderToStringOptions = PrerenderOptions;
+
+async function renderToString(
+  vNode: ReactNode,
+  options: RenderToReadableStreamOptions
+): Promise<string> {
+  return readableStreamToString((await prerender(vNode, options)).prelude);
+}
+
+async function readableStreamToString(readableStream: ReadableStream) {
+  let result = '';
+  const textDecoder = new TextDecoder();
+  for await (const chunk of readableStream) {
+    result += textDecoder.decode(chunk, { stream: true });
+  }
+  result += textDecoder.decode(); // flush end character
+  return result;
+}
+
 export {
-  prerender,
-  PrerenderOptions,
+  renderToString,
+  RenderToStringOptions,
   renderToReadableStream,
   RenderToReadableStreamOptions,
   ReactDOMServerReadableStream,
